Added a bunch of new things
- Added food.
- Added wall collision
- Added body segments

- Problems I found -

It looks like everytime I eat an apple, the segment get inside the head. So I added "segments[0].color("white")"(line 114) in the segments lines of code. That way, the first segment that gets created inside the head is white (the original head color) and the other segments (the body) are black.

Another problem is that when I start the game, the first apple is always on this coordinates (0, 100) and I don't know a way to put it in a random place on the canvas when I start a new game.


Sorry if I'm bad at this. It's my first time programming and using this website. I'm trying to learn and integrate in this society. Btw, I'm only 15.